requirements.txt
# dependencia de projeto
Django==5.0.4
djangorestframework==3.15.1
django-filter==25.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
pillow==10.3.0


# dependencia de ambiente
python-decouple==3.8

gunicorn==22.0.0
uvicorn==0.29.0